backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==0.58.2
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1
docstring_parser==0.17.0
google-api-core==2.26.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.121.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.45.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jwt==1.4.0
MarkupSafe==3.0.3
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.5
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
